reparated code into service (execution) and factory (creation) + renamed ILoadingOptions -> ITdLoadingConfig
